Make batched state fallback to initial state after flushing updates
Because batched state will be called when rendering is done and the updates
will be flushed, it is not necessary to call `setState` on unmounting a component
with props which will turn "off" whatever batched state is being used to control.
Instead, set initial state of the batched state to the "off state", the render
function can then always turn on the state, and if components haven't been mounted
then the setState will not be called and it will default to the off state.